requirements.txt
pandas>=2.0
pyarrow>=14.0
pyyaml>=6.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
rank-bm25>=0.2.2
